Use a ChainedScheduler per Destination
When reading persistent pending events, do not schedule them right away.
Instead put these events into a ChainedScheduler that will run in the
threadpool for which the push is Destined. This puts the work of
scheduling where it belongs so that each Destination has less impact on
the other Destinations. This also lowers the priority of the pending
events with respect to newer user generated events. This priority
lowering prevents a potentially huge backlog of stored pending events
from starving current ongoing server activity. This is valuable since
when there is a huge backlog of pending events, the backlog likely came
from a 'replication start --all' command which tends to be lower
priority than actual user generated ref updates.

On our servers, a backlog of events created by a 'replication start
--all' tends to last around 45minutes. We have automated running this
command every time a master shuts down, and also if the entire cluster
ever starts back up. This is in order to avoid any long term, but very
unlikely, loss of updates which did not get recorded during shutdown.
This automated action serves the function of 'startup replication' in a
multi-master environment, which unfortunately the design of the
persistent event store does not completely eliminate the need for.

Since these replication event surges happen during our rolling updates,
they partially impact the ability of the rolling update to minimize user
service interruptions. The impact of most of the user replication events
being delayed by up to 45minutes is much larger than the impact of a
very unlikely missed shutdown event being delayed by a likely very small
amount due to this change. Thus this priority lowering is a trade-off
that can have a huge service improvement for us.

This change also has the impact to potentially reduce replication
replaying work in two different ways. Both the improvements relate to
the splitting of the reading and the scheduling into two separate parts.

The first improvement is that different ref updates to the same URIs are
now much more likely to be combined into a single push. Previously,
events were scheduled as they were read, which meant that any delay in
scheduling would delay the next read. As scheduling is very slow on
server startup, the scheduling delay meant that a large portion (usually
around 50%) of the replication events would have already completed
before the scheduling completed. Since previously URI consolidation
could only happen after scheduling, there was an approximately 50%
chance that the consolidation would not happen, i.e. it would not happen
when the to-consolidate-with push had already completed. However since
there is now a queue of updates to be scheduled that gets loaded in a
very short period of time, that can consolidate pushes to the same URI,
most updates should get consolidated.

The second work reduction comes from the additional fact that this
change also no longer delays configuration reloads while replaying
events. As events will be stored on disk and will replayed again anyway
after a reload, this should be OK. Since the queue of scheduled events
will generally be shorter with this approach (as they will sit in a
'toSchedule' queue most of their waiting time before being scheduled),
the events which get re-replayed after a reload should be much smaller,
likely no more than the size of each threadpool +1, this can result in a
huge savings by de-duplicating work for reloads and thus making them
more useful.

Change-Id: Idf7019ed17ba993c2a3cedd22f379ca487d2dbf2