Fix testing and replication of project creation/deletion
Make sure project creation and deletion are correctly managed
and tested, making more assertion on the whole set of branches
being replicated and not just the bare repo showing up.

With the extra test coverage, fix a regression where the project
deletion was not replicated to the remote nodes because its associated
project state was missing locally.

When replicating deletions, the project state is not present as
the replication plugin is receiving the deletions ex-post
and therefore do not have access to the project anymore.

Bug: Issue 12829
Bug: Issue 12806
Change-Id: I0fdc1e73390c2abd3e40d2a02fd7e4ce7f20bb67