Fix replication of project deletion
Fix a regression where the project deletion was not propagated
to the remote nodes because its associated project state was
missing from the project cache.

When replicating deletions, the project state is not present as
the replication plugin is receiving the deletions ex-post
and therefore do not have access to the project anymore.

The associated checks for project visibility and read only state
are not valid for project deletions.

Bug: Issue 12806
Change-Id: I7a9ac01b01d5dd40b8bf0c4d3347256f430329ac