Fix replication of project deletion
Fix a regression where the project deletion was not propagated
to the remote nodes because its associated project state was
missing from the project cache.

When replicating deletions, the project state is not present
as the replication plugin is receiving the deletions ex-post
and therefore do not have access to the project anymore.

The associated checks for project visibility and read only state
are not valid for project deletions.

Change-Id: Ie7e6e425f476a720bb54f3e3afaf801601ec6717
Bug: Issue 12806