Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > Thanks, Davido, for this change. Looking forward to see JSch removed :-)
> > Can we add one test for demonstrating that this change doesn't break anything?
> 
> The corresponding gerrit core change is running dozen (or hundreds?) of unit and intergrations tests using Apache MINA SSH client. Isn't this enough?

Apologies, I did not express myself clearly. I was asking for a test that the replication plugin works with this change.

> The real test would be:
> 
> * Set up gerrit site from the gerrit core change: [1]
> * Set-up replication plugin
> * Configure couple of destinations
> * Confirm that the replication works as expected

Yes, we do already the ReplicationIT suite that simulates this by using the same Gerrit server as source and destination and that works. I am wondering if we can extend that suite.

Hope this clarifies.

Patch-set: 2