Update patch set 1
Patch Set 1:

Why not sharding also the other directories? Are you expecting to have a lot more events in waiting rather than running? If yes, how is the system going to scale if you always have a lot more to do compared to what you are capable of doing?

Patch-set: 1