Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Why not sharding also the other directories? Are you expecting to have a lot more events in waiting rather than running?
> 
> Yes
> 
> > If yes, how is the system going to scale if you always have a lot more to do compared to what you are capable of doing?
> 
> I don't think "always" is part of the assumption here. There can be times where there is a build up of many things to do (like start up replication) and then those drain over time.

If the node is down you cannot possibly accumulate events in the waiting queue.
Can you share some numbers of what was the increase in performance of sharded vs. non-shareded, with regards to the overall E2E replication times?

In my experience, sharding makes sense when you have tens of thousands of files, which means tens of thousands of replication push in the backlog. In my experience you could have a "backlog" of replication events of course, but, if your system is coping with the load, it should be limited to a few hundreds.

Some maths for a medium-sized repo where a push would take, let's say 100s. For 1000 events then:
- 1000 events x 100 sec = 100k sec = 27h (if serialised) and with a 10x parallel replication threads => 2.7h

If you guys really have tens of thousands of events in the backlog that sharding on the filesystem of the events file becomes a thing, then it means you may have 10k or 100k of events in the backlog !

Can you share some numbers?

Patch-set: 1