Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Verified+1
> > 
> > Verified in collabnet's replication integration tests.
> 
> That's cool, how can we make sure that won't break in the future? (merge to stable-3.2 and master)
> Would this be possible to add as test-case to ReplicationIT?

I think the risk of it breaking is low. It's a no-op change for the default use case and is only used by collabnet.

The only reason it's not already on stable-3.1 is because I didn't have time to resolve the conflict on the last merge up from stable-3.0 and we (collabnet) were not using 3.1 then anyway.

Since we're using it now, and it's covered by our internal integration tests, if it does break we will notice.

Patch-set: 2