Re-add getting URI lock before running tasks
This reverts commit 09ea614701fd0c4f568bccfcb967d120960afec1 and
squashes in the bug fix for potential event loss during external
collisions and avoiding using magic values for external push ids.

Previously when a URI was being replicated to by a node (node A) in a
shared storage multi-master setup without a replication distributor
enabled, a new push event generated by another node (node B) could be
lost. This loss would occur if node B attempted to push to the same URI
as node A, but for a different newer push event generated locally while
the push was still ongoing on node A. In this case, node B would be
denied the URI lock and behave as if the push were already completed. As
this push is for a different newer event, dropping this push event would
result in it being lost.

This bug would occur because the code treated externally denied
events, which could be created either by external completions or
external in-flight operations, the same. Now, in the in case of
externally in-flight pushes, treat them similar to local in-flight
collisions, and reschedule them.

Additionally add tests to simulate these two different scenarios.

When a replication runway is externally denied, use a local const
self-explaining value rather than the hardcoded magic values.

Change-Id: I5dd27968e8af9744547d678a93242f6e28b20e6e
Bug: Issue 12933