TasksStorage: Replace delete() with start()+finish() in tests
During the merge of stable-3.0, the delete() method was re-added [1]
to fix a test that relied upon it. Instead, that test should have been
updated to use the start() and finish() methods that are part of the
public API for ReplicationTasksStorage.

To facilitate that update, modify the start(), reset(), and finish()
methods to not require a PushOne object. These methods can instead
use a new interface that PushOne already can implement without
functional changes. This interface avoids duplicate code in
ReplicationTasksStorage.

[1] https://gerrit-review.googlesource.com/c/plugins/replication/+/273660/-1..4/src/main/java/com/googlesource/gerrit/plugins/replication/ReplicationTasksStorage.java

Change-Id: Id82d2ec3125075832134c8dbe56b342e5a874bbc