TasksStorage: Remove delete() only used by tests
In general we shouldn't have methods that provide behavior only for
tests. Callers of delete() in pre-3.1 code should move to start() and
finish() in 3.1. To facilitate that, update start(), reset(), and
finish() methods to not require a PushOne object. These methods can
instead use a new interface that PushOne already can implement without
functional changes.

Change-Id: Id82d2ec3125075832134c8dbe56b342e5a874bbc