ReplicationTasksStorage: Add multi-master unit tests
These tests examine the replication scenarios under multi-master
setup making use of the api calls present in ReplicationTasksStorage
class similarly as done in single master setup.

Eventhough replication in multi-master setup is a planned feature
for v3.3 we should aim for its compatibility and these tests ensure
that compatibility is not broken since we are testing something that
already works.

Change-Id: I375b731829f3c0640d3a7a98635e1e5c526908ca