Introduce hashCode/equals in ReplicateRefUpdate
ReplicateRefUpdate is used extensively in ReplicationTasksStorageTest.
Allow using it with the collections assertions so that reading and writing
tests becomes easier and the failures more self-explanatory.

Allow also in the future commits to leverate the beans properties
and being used flawlessly in collections.

This is a perparation work for introducing then the AutoValue in
master for representing all the beans in the replication plugin.

Change-Id: Ib1e68a520043869f4af8a7ece776cdb9c48c5a62