Introduce hashCode/equals in ReplicateRefUpdate
ReplicateRefUpdate is used extensively in ReplicationTasksStorageTest.
Allow using it with the collections assertions so that reading and writing
tests become more natural and the failures more self-explanatory.

Use the new capability to use proper Set<ReplicateRefUpdate> in the
replication plugin, which represents more adequately nature and
the granularity of the ref-updates.

The introduction of a proper AutoValue is planned to be implemented on the
master branch.

Change-Id: Ib1e68a520043869f4af8a7ece776cdb9c48c5a62