Reproduce multi-primary replication failing scenario
Under a multi-primary replication setup, when two different nodes try
to replicate to same uri and ref, with specific sequence of events,
there is possibility for one of the nodes to lose the persisted event
to be replicated.

In order to not lose events, it is crucial to have some unique ID
associated with events created by specific node.

Bug: Issue 13441
Change-Id: Ie35550527f18c3eda1b8b81b7a863d634ef5755a