Move storage portion of replicateBranchDeletion ITs into ReplicationStorageIT
All other ITs did this split on stable-2.16, so this is only updating
the new replicateBranchDeletion tests that were added in stable-3.0.

The e2e check for if the branch is removed or not stays in ReplicationIT
and the check that a task is created in storage when the branch delete
API is invoked moves to ReplicationStorageIT.

This split allows the best practices for verifying e2e and storage to be
applied independently.

Change-Id: Iec7ee090bd614e3442b1f9cb454437c9e05290be