ReplicationIT: refactor shouldMatch* to not rely on storage state
Tests in ReplicationIT should verify e2e behavior (i.e. does a remote
get a replication update, or not) only. All verification of storage
correctness is already covered through tests added to
ReplicationStorageIT in Iaa14a74 so we are not reducing test coverage.

Also removes the disableDeleteForTesting flag as there are no users of
it now.

Change-Id: I4ae15d0301a308a12cbca3684915e89ca421e02f