Add new ReplicationStorageIT tests
Refactor tests in ReplicationIT to not rely on storage state to verify
e2e behavior and move+create tests in ReplicationStorageIT dedicated to
verifying storage correctness.

Tests in ReplicationStorageIT utilize long replication delays so that
tasks in storage are in a predictable state.

Also removes the disableDeleteForTesting flag as there are no users of
it now.

Improved code sharing between ReplicationIT and ReplicationStorageIT
will be addressed in a later commit.

Change-Id: I41179c20a10354953cff3628368dfd5f910cc940