ReplicationFanoutIT: Split shouldReplicateNewBranch tests
Split these into storage-based and e2e tests so that the storage tests
can be reliably verified through use of replicationDelay large enough
that task state on disk doesn't change during the tests.

Keep them all in ReplicationFanoutIT for now since the setup for these
tests is unique to that class.

Also remove the unnecessary cleanup of tasks.

Change-Id: I36e0a4affe1f5d1330ea27a496fd8ba295176763