Fix replication to retry on lock errors
Versions of Git released since 2014 changed the message returned
from "failed to lock" to "failed to update ref" when the remote
is unable to lock a ref. Refer Git commit:

https://github.com/git/git/commit/6629ea2d4a5faa0a84367f6d4aedba53cb0f26b4

Config 'lockErrorMaxRetries' is not removed as part of this change
as folks who have it configured currently don't run into unexpected
behavior with retries when they upgrade to a newer version of the
plugin. Also, the "failed to lock" check is not removed for folks
still using a version of Git older than 2014.

Change-Id: I9b3b15bebd55df30cbee50a0e0c2190d04f2f443