Avoid starting tasks which are not present under waiting storage area
This change only starts the replication tasks which are present
under the ../waiting directory and ignores any extra refs which
it was orginally intended to push. This will reduce the duplicate
work because the absence of tasks under ../waiting directory
implies that it has already been started/completed by other
nodes.

Change-Id: Ifbb7018ec1d960015626c089a4dadf6b0247d278