Avoid starting tasks which are not present under waiting storage area
This change starts the replication tasks which are only present
under the ../waiting directory and drops off any extra refs which
it was orginally intended to push. This would reduce the repeated
workload because the absence of tasks under ../waiting directory
implies that it has already been scheduled/completed by other
nodes.

Change-Id: Ifbb7018ec1d960015626c089a4dadf6b0247d278