Remove dependency on commons-io library
This dependency was added in Id12780948a4 to support remoteNameStyle
"basenameOnly". The only reason for this dependency is to translate
project name from "foo/bar/myrepo" to myrepo.

It seems to be overkill to add 169 KB to the plugin distribution
for one single method.

Another disadvantage is that the version of common-io library used
for this is 2.2 from 2012. If gerrit installation site is using some
other plugins in addition to replication plugin, then it can easily
lead to classpath collision, when different versions of commons-io
libraries are included as transitive dependencies of different plugins.

To rectify, use a replacement method from guava library.

Change-Id: Id254dc38831832a9855bd204e4c2129ec64b88ae