Prevent concurrent event replaying
If the events distributor is enabled and startup events replaying takes
longer than the distributor delay, the two threads will be replaying
events concurrently. This contention is unhelpful as we could be
rescheduling tasks that we've already scheduled.

The distributor itself is already prevented from running concurrently by
ScheduledThreadPoolExecutor's scheduleWithFixedDelay() behavior.

Change-Id: Idce111dfb54a88ab6ace0f52e70b61a711395855