Notify about project deletion status
Notify the scheduling and the result of the execution of a project
deletion on remote targets.

This is useful for consumers who are interested in understanding not
just when a project deletion is initiated, but also when it is
completed.

The same event notification was applied to ref updates through the
propagation of RefReplicationScheduled and RefReplicationDone events,
but it was never applied to project deletions.

Feature: Issue 13894
Change-Id: I9b8197e67f4eddcc51c408c2db4c5991487a3d5e