Fix cases where replication is waiting indefinitely
In change [1], we started to avoid starting replication tasks which are
not present in ../waiting dir. But the replication commands were waiting
due to state not being updated for the work which is avoided. This change
marks the refs which are neglected as not attempted and in turn notifies
that update to ReplicationState object.

[1] Ifbb7018ec1d960015626c089a4dadf6b0247d278

Change-Id: Iff31d62ebbbfb88754b43b60344e05dd4b0a1f6d