events: Add targetUri to deprecate targetNode
RefReplicatedEvent and ReplicationScheduledEvent have a targetNode field
that is populated using a method intended for human-readable SSH command
output. Deprecate that field and add a new targetUri that is the ASCII
string representation of the URI. This new field is more suitably named
and populated for consumers.

This will not break existing users of targetNode until a later change
removes that field. The only known user of the targetNode field is the
Jenkins Gerrit Trigger plugin [1][2].

[1] https://github.com/jenkinsci/gerrit-trigger-plugin
[2] https://github.com/sonyxperiadev/gerrit-events/blob/d15f38adc7ea90a98486dbe0df01d31335aaa3af/src/main/java/com/sonymobile/tools/gerrit/gerritevents/dto/events/RefReplicated.java

Change-Id: If07f7103a1f9cf9e49e5eef4c91e9de1b5e46963