events: Add targetUri
RefReplicatedEvent and ReplicationScheduledEvent have a targetNode field
that is populated using a method intended for human-readable SSH command
output. Deprecate that field and add a new targetUri that is the ASCII
string representation of the complete URI.

This will not break existing users of targetNode until a later change
removes that field. There are no known public users of targetNode.

Change-Id: If07f7103a1f9cf9e49e5eef4c91e9de1b5e46963