Call retryDone() when giving up after lock failures
Previously when giving up after retrying due to too many lock failures,
a 'replication start --wait' command would wait indefinitely if it was
waiting on the push that gave up. Fix this by calling retryDone() after
giving up which will trigger the ReplicationStatus to reflect a failure
allowing the waiting to complete.

Change-Id: I0debade83612eb7ce51bab0191ab99464a6e7cd3