Add possibility to split large pushes to remote into batches
Allow to split push with a large number of refs into smaller,
size-configurable batches to:
1. allow task cancellation - after each batch gets pushed operation can
   be checked if it was requested to stop
2. set push operation timeout more predictably, with this parameter set
   one can more definitely specify when timeout should occur
3. successfully push to remotes like GitHub, GitLab that enforce limits
   on the number of refs that could be accepted in a single push operation
4. ability to reduce the `price-to-pay` for large push failues e.g.
   without batching when push for large amount of refs fails all of them
   need to be pushed again. Whereas when batches are enabled only failed
   ones need to be pushed again.

One can specify the batch size with 'gerrit.pushBatchSize' parameter of
'replication.config'. By default `0` is assumed which means that there
is no limitation on number of refs in a single push operation (current
behaviour is preserved). Note that negative values are treated as `0`.

Bug: Issue 13216
Change-Id: I1514855cbb5ac0ea78397a17eb1c5926bf07a4a3