Ensure that meta ref is pushed together with change ref
Especially for large projects it could happen that a replication
push of a change ref is triggered before the meta ref of this
change is pushed. With large replication delays that could prohibit
clients to fetch that change, when reacting on ref-replicated events.

Now the respective meta ref is added to each replication push of a
change  ref. If the meta ref is already up to date on the remote,
this is a no op. If it would have been in the same push operation,
nothing changes. If it does not yet exist, the push operation will
be rescheduled.

Bug: Issue 15194
Change-Id: Ic10af9b75cda494310e69399a331793906ca05f7