Documentation: remoteNameStyle might result in a repo name clashes
When using the remoteNameStyle in the replication.config settings
the Gerrit admin needs to warned on the risks of using sytles
that may be ambiguous and associated two source repositories to
the same target repository on the replica.

Example: set remoteNameStyle to basenameOnly

/foo/my-repo.git => pushed to my-repo
/bar/my-repo.git => pushed to my-repo

When two commits are pushed to the same branch on the two repos
/foo/my-repo.git and /bar/my-repo.git, the replication plugin
would push them to the same target repo my-repo, causing clashes
and losing commits (depending which one is pushed first).

The risk needs to be highlighted so that the Gerrit admin
can check that univocity of the mapping is respected.

Change-Id: Iba42907bceb8d1c27d739f3b0cded4a1d7400686
Bug: Issue 15315