Documentation: make explicit that remoteNameStyle is for non-Gerrit repos
The documentation already specify non-Gerrit examples of use
of remoteNameStyle; however, it does not say that if the remote
repository is backed by Gerrit, the *ONLY* supported option is
"slashes", otherwise the consequences could be catastrophic.

Two Gerrit servers (e.g. primary and replica) need to have
full alignment of repository names, as they are also referenced
in the inherited ACLs. Having a repository name mapping, may
disrupt the ACLs evaluation and make the remote Gerrit replica
unusable.

Bug: Issue 15318
Change-Id: I4d9447a4d0366a98037470c0cceda36f7a1b8a25