Adapt to change of the ref updated event
The ref updated event was changed to contain all refs that were
changed in a single operation.

This change adapts to the change in Gerrit core. This also fixes
the issue that a change-ref could be pushed before the associated
meta-ref. In such cases, clients trying to fetch the change on the
receiving side failed to do so until the meta-ref was pushed as well.

Bug: Issue 15194
Change-Id: Iddf831e0c6603c0124ea6ab03494737d7b03de28