Replicate refs created in a BatchRefUpdate together
With a recent change it was introduced that internal ref-updated
events would contain all refs that were updated in a single command.
This was not yet used by the replication plugin to ensure that
refs that are dependent on each other are guaranteed to be replicated
in the same push.

Now, each replication task contains a set of one or more refs instead
of just one. This allows to define tasks that take care of multiple
refs that were updated by a single command.

Depends-On: https://gerrit-review.googlesource.com/c/gerrit/+/328228
Change-Id: Ic0534d03dee6759ba4cb6b19ae87fd75269b437b