[DO NOT MERGE] Expose fragility of repository creation logic
This failing test exposes a problem where the creation of remote
repositories is based on a fragile algorithm that matches substrings of
error messages.

For example, if the remote throws an exception containing
'Git repository not found' a repository is created, whilst if the
exception contains 'not found', the repository is NOT created.

The current list of errors[1] is arbitrary and incomplete and a new
solution should be proposed.

Bug: Issue 15538
[1] https://gerrit.googlesource.com/plugins/replication/+/refs/tags/v3.5.0.1/src/main/java/com/googlesource/gerrit/plugins/replication/PushOne.java#439

Change-Id: I22ac09d1815e9ed2146915018dae8cb3d4e3bfa3