Introduce hasProject in AdminApi
Stop guessing whether the remote destination requires the creation of a
repository by just looking at the string in the caught RemoteException.

This is error-prone and the list of strings to match against can only be
an arbitrary subset of all possible strings.

Instead, actively check whether the remote destination actually has a
repository by asking the remote directly.

Bug: Issue 15538
Change-Id: I8fe4495517886ed350e83af1e8f758376e396e83