Fix creation of missing repository
On gerrithub, we are experiencing failure in remote repository creation,
because the error thrown by the remote is: "not found.", which is not
among the current list of expected messages.

As a fix for stable branches, add the additional error message to the list.

A more stable fix for this has been proposed [1] for master, so that the
detection of repository creation is more robust.

[1]https://gerrit-review.googlesource.com/c/plugins/replication/+/329028

Bug: Issue 15538
Change-Id: I79d6f5c1028e301130615ecedd6083c7adc9d6d1