Ensure states are updated for canceled replication tasks
When using 'replication start --wait' we need
ReplicationState.waitForReplication() to return when the task we're
waiting on has been canceled, either through an admin action or because
the replication distributor determined another node already completed
it.

Add a couple tests for PushAll that confirm this behavior was previously
broken and is fixed now.

Change-Id: I36320ae079af5d7673e05d20ddc94b42a9b04347