Do not retry replication when local repository not found
When many repositories get created and deleted in a quick succession it
may happen that a repository gets deleted before its replication task
starts. Such replication task will keep retrying, possibly indefinitely,
but has no chance to succeed.

When handling the RepositoryNotFoundException in PushOne, set the
retrying flag to false. This ensures that this replication task is
not retried and gets finished.

Change-Id: Ia55c5ec1c961f4c2aec9ecee8056f22b436e9fda