Allow to finish waiting tasks for non-existing repos without scheduling them
A waiting task for a non-existing repository cannot be (re)scheduled
because there is a check for repository existence in that code-path.
However, such a task would remain in the waiting queue and rescheduling
would be tried again and again without a chance to get finished as the
only way for a task to get finished was to run it.

This change allows to finish a waiting task when its repository doesn't
exist. The ReplicationTasksStorage now tries to delete task file(s)
from both running and waiting directories.

Change-Id: Ibbdd5023e2a008484215da02403c9935d21fbf13